Copy Sharebox image resources over to app bundle.
To allow customization of icons and the sharebox background, icons and
sharebox are copied over to the "resource folder of the NitroShare app
bundle. Now simply "Show Package Contents" of nitroshare.app and
navigate to Contents -> Resources , after the first launch of
NitroShare, in order to replace or edit the icons, deleting the changed
icons will recreate the originals on next launch.